Initial checkin of CmsSystemInfo object.
The CmsSystemInfo object is a container for System - wide information
that is "read only" in nature. Example is the web application name, 
or the path of the OpenCms installation. It's similar in nature to the Java
"System" object, only that is not a property list and all information is read 
only.