Solr: Restricted name based con_locale detection to available locales.
Files with suffix _xy or xy_zab were indexed under locale xy or xy_zab
before. Now the suffixes are only taken into account when they are
available locales for the resource, i.e., mentioned in the (inherited)
locales-available property (or, if not set, the default locale, which
would not influence the con_locales value at all).