content editor save action: do not wait for index updates.
Before it was waited up to 30 seconds for the offline indexes to be
updated on a save action. This was confusing for the editor.
Now it might happen more often that indexes are not up to date after
editing (e.g., after mapping a content element to the Title property,
the content will be found under the old title for some seconds, before
the index is updated).