Site Manager app: some improvements
Bundle name for resolving macro-method is defined as static String in
CmsSiteManager (currently ="siteMacroBundle"). A bundle with this name
(plus "_en" or similar) is supposed to be in the .macros folder for
sites used as a template-site for new sites.
UI-changes: Size of text fields corrected, for adding new sites an own
tab was introduced.