Added Solr update processor for copying and manipulate (via regex) a value from one field to another
The processor has to be configured to take action.

The original use case is for a document center where multiple language
versions of a document could be grouped by generating a
(locale-independent) path version for each of them. E.g. convert the
original paths "/doc_en.pdf" and "/doc_de.pdf" (from field "path") both
to "/doc.pdf" (and store that information in field "path_remove_locale")
and collapse/expand on the "path_remove_locale" field when querying.