Basic library updates. The project depends on a large number of libraries updated to a new version; As the upgrade to Servlet3.1, so upgrade the Java version from 1.7 to 1.8.
Because it is a base class library, so a lot of Java classes changed.
But mainly not some complex business, so the logic is very simple, there
should be no Bug generated. I also did a test:
Use "gradle bindist" BUILD SUCCESSFUL;
The deployed WAR is deployed to TOMCAT and is working.

Conflicts:
	dependencies.gradle
	src/org/opencms/configuration/CmsImportExportConfiguration.java
	src/org/opencms/configuration/CmsSystemConfiguration.java
	src/org/opencms/module/CmsModuleXmlHandler.java
	src/org/opencms/util/CmsHtmlTagRemoveFactory.java