Read search directory also from system property
The tests can fail if running the data of the solr indexes on a
NFS volume (See https://github.com/alkacon/opencms-core/issues/585).

This commit allows to configure the OpenCms search directory (under
which the data indexes are stored) with a system property
`opencms.search.directory` that overrides the value set in
`opencms-search.xml:/opencms/search/directory`.