Search indexer: removed error forwarding of updateResource and all subsequent throw declarations.
The error was misleading in the way that it could only be trigged by
runtime exceptions. Other exceptions were catched already. Hence, it did
not correctly indicate that indexing one document failed.
Moreover, if the error was thrown for one buggy the whole indexing
process was quit - i.e., the healthy documents where not updated as
well.