Search: Adjusted document type configuration. CONFIG HAS TO BE ADJUSTED.
The document factories used for indexing, as configured in
opencms-search.xml are now determined differently.

See the changes in webapp/WEB-INF/config/opencms-search.xml for
necessary adjustments in the configuration.

In essence we now allow for extraction of the same resource with
different document factories in different indexes. E.g., xmlcontent
resources are extracted differently in plain lucene and solr indexes
(configured by document types xmlcontent and xmlcontent-solr).

Moreover, it now becomes possible to add document types for specific xml
contents that allow to extract them differently.

For the mechanism, on how a document factory is selected see the javadoc
for org.opencms.search.CmsSearchManager#getDocumentTypeKeys(String
resourceType, String mimeType).