Fixed broken xml content mapping test case.
(The previous change to mappings caused testMappingsWithManyLocales to
fail, because of a mapping to the locale property. This mapping was
actually tested in testMappings, but it was redundant since the same
type of mapping is also tested for the NavImage property, so we removed
the test code for the locale property to prevent the test case
failures).