Fix: Prevent loading all pages of results where more than maximally searched results are present.
OpenCms does search at maximum for 10000 items for the gallery search.
But if e.g., you have 19000 items and only permissions on 10, OpenCms
would have loaded all pages until potentially the 19000 items could have
been shown. We omit this by stopping the reload if a page is loaded with
less than the requested items. This is sufficient to know that all
searchable items are already found.