Refactoring of ListManager
Moved all parsing and search relevant information to
org.opencms.jsp.search.config.parser.simplesearch, placed nested static
classes into seperate files and seperated parsing of the config content
from the bean holding the parsed content.

The basic refactoring idea is to see the <cms:simplesearch>-tag as the
main feature and the app just as an "add-on" where the search feature
itself is not dependent on.