Updated xalan to make <x:transform> JSTL tags work again.
The previous version of Apache Xalan had a bug that caused attributes to be deleted when using XSLT transformations in JSPs. We have to use Xalan and can't use the JVM XSLT implementation because we depend on features specific to Xalan. The original Xalan project is no longer maintained, but there is an 'inofficial' bugfix version as part of the Apache ServiceMix project, which is available on Maven Central, which this commit adds as a dependency.