Changed behavior of FlexCache 'timeout' directive so that it takes expiration dates into account.
Normally, the expiration date of a FlexCache entry is computed based on
the expiration dates of the resources read while creating it. But,
before this change, if a timeout cache directive was used, then the
resource expiration dates would be completely ignored. This caused
errors with container pages if the expiration date of a container
element happened earlier than the timeout of the cache entry of the
formatter/template containing the <cms:container> tag as part of which
the elements were rendered.