Fixed problem with publishing sibling relations (github issue #805).
After changing the links in an XML content and publishing it without its
siblings, the siblings would previously show up as unchanged in the
Explorer despite the fact that their relations have been changed to
reflect the changed content. This has been changed so that both
unpublished siblings will have the 'changed' state and need to be
published.